split out the auth tests ensure we respond with the correct exception when user name / password is incorrect add tests for username / password errors
