+++ src/main/java/io/mapsmessaging/network/protocol/impl/mqtt5/listeners/AuthListener5.java
@@ -19,31 +19,35 @@
 package io.mapsmessaging.network.protocol.impl.mqtt5.listeners;
 
 import io.mapsmessaging.api.Session;
+import io.mapsmessaging.logging.ServerLogMessages;
 import io.mapsmessaging.network.io.EndPoint;
 import io.mapsmessaging.network.protocol.ProtocolImpl;
 import io.mapsmessaging.network.protocol.impl.mqtt.packet.MalformedException;
 import io.mapsmessaging.network.protocol.impl.mqtt5.AuthenticationContext;
 import io.mapsmessaging.network.protocol.impl.mqtt5.MQTT5Protocol;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.Auth5;
+import io.mapsmessaging.network.protocol.impl.mqtt5.packet.ConnAck5;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.Connect5;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.MQTTPacket5;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.StatusCode;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.properties.AuthenticationData;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.properties.AuthenticationMethod;
 import io.mapsmessaging.network.protocol.impl.mqtt5.packet.properties.MessagePropertyFactory;
+import io.mapsmessaging.utilities.scheduler.SimpleTaskScheduler;
 import java.io.IOException;
+import java.util.concurrent.TimeUnit;
 
 public class AuthListener5 extends PacketListener5 {
 
   @Override
   public MQTTPacket5 handlePacket(MQTTPacket5 mqttPacket, Session session, EndPoint endPoint, ProtocolImpl protocol) throws MalformedException {
     // Need to push this until we have finished our auth
-    AuthenticationMethod authMethod = (AuthenticationMethod) mqttPacket.getProperties().get(MessagePropertyFactory.AUTHENTICATION_METHOD);
     AuthenticationContext context= null;
     if (mqttPacket instanceof Connect5) {
       try {
         MQTT5Protocol mqtt5Protocol = ((MQTT5Protocol) protocol);
         if(mqtt5Protocol.getAuthenticationContext() != null) {
+          AuthenticationMethod authMethod = (AuthenticationMethod) mqttPacket.getProperties().get(MessagePropertyFactory.AUTHENTICATION_METHOD);
           String serverConfig = mqtt5Protocol.getAuthenticationContext().getAuthMethod();
           String clientConfig = authMethod.getAuthenticationMethod();
           if (!serverConfig.equalsIgnoreCase(clientConfig)) {
@@ -52,6 +56,7 @@ public class AuthListener5 extends PacketListener5 {
           context = mqtt5Protocol.getAuthenticationContext();
           mqttPacket.getProperties().remove(MessagePropertyFactory.AUTHENTICATION_METHOD);
           context.setConnectMsg(mqttPacket);
+          context.setAuthenticationMethod(authMethod);
         }
       } catch (IOException e) {
         throw new MalformedException("Exception raised creating Authentication Server", e);
@@ -63,30 +68,47 @@ public class AuthListener5 extends PacketListener5 {
     //
     // OK we have done the initialization above, lets process the auth packet
     if (context != null) {
-      AuthenticationData clientData = (AuthenticationData) mqttPacket.getProperties().get(MessagePropertyFactory.AUTHENTICATION_DATA);
-      try {
-        byte[] clientChallenge = context.evaluateResponse(clientData.getAuthenticationData());
-        if(clientChallenge != null && clientChallenge.length > 0){
-          byte state = StatusCode.CONTINUE_AUTHENTICATION.getValue();
-          if(context.isComplete()){
-            state = StatusCode.SUCCESS.getValue();
-          }
-          Auth5 auth5 = new Auth5(state, context.getAuthMethod(), clientChallenge);
-          ((MQTT5Protocol) protocol).writeFrame(auth5);
-        }
-        if (context.isComplete()) {
-          ((MQTT5Protocol) protocol).setAuthenticationContext(null);
-          MQTTPacket5 initial = context.getConnectMsg();
-          return ((MQTT5Protocol) protocol).getPacketListenerFactory().getListener(initial.getControlPacketId()).handlePacket(initial, session, endPoint, protocol);
-        }
-        return null;
-      } catch (IOException e) {
-        throw new MalformedException("Exception raised in processing Auth challenge", e);
-      }
+      return handleAuth(context, protocol, mqttPacket, session, endPoint);
     } else {
       throw new MalformedException("Expected Authentication Context but none found");
     }
   }
 
+  private MQTTPacket5 handleAuth(AuthenticationContext context, ProtocolImpl protocol,MQTTPacket5 mqttPacket, Session session, EndPoint endPoint) throws MalformedException {
+      AuthenticationData clientData = (AuthenticationData) mqttPacket.getProperties().get(MessagePropertyFactory.AUTHENTICATION_DATA);
+      byte[] clientChallenge;
+      try {
+        clientChallenge = context.evaluateResponse(clientData.getAuthenticationData());
+      } catch (IOException e) {
+        // Auth has failed, so we need to send a ConnAck with a reject and close the connection
+        ConnAck5 connAck = new ConnAck5();
+        connAck.setStatusCode(StatusCode.BAD_USERNAME_PASSWORD); // MQTT Standard
+        connAck.getProperties().add(context.getAuthenticationMethod());
+        connAck.setCallback(() -> SimpleTaskScheduler.getInstance().schedule(() -> {
+          try {
+            protocol.close();
+          } catch (IOException e1) {
+            logger.log(ServerLogMessages.END_POINT_CLOSE_EXCEPTION, e1);
+          }
+        }, 100, TimeUnit.MILLISECONDS));
+        return connAck;
+      }
 
+      // Auth is ok to continue, so check to see if it has completed or needs more information
+
+      if(clientChallenge != null && clientChallenge.length > 0){
+        byte state = StatusCode.CONTINUE_AUTHENTICATION.getValue();
+        if(context.isComplete()){
+          state = StatusCode.SUCCESS.getValue();
+        }
+        Auth5 auth5 = new Auth5(state, context.getAuthMethod(), clientChallenge);
+        ((MQTT5Protocol) protocol).writeFrame(auth5);
+      }
+      if (context.isComplete()) {
+        ((MQTT5Protocol) protocol).setAuthenticationContext(null);
+        MQTTPacket5 initial = context.getConnectMsg();
+        return ((MQTT5Protocol) protocol).getPacketListenerFactory().getListener(initial.getControlPacketId()).handlePacket(initial, session, endPoint, protocol);
+      }
+      return null;
+  }
 }
